Restructured as per the lecture and the Flask guide
