+++ requirements.txt
@@ -1,8 +1,13 @@
+alembic==1.18.4
 blinker==1.9.0
 click==8.3.2
 Flask==3.1.3
+Flask-Migrate==4.1.0
+Flask-SQLAlchemy==3.1.1
 itsdangerous==2.2.0
 Jinja2==3.1.6
+Mako==1.3.12
 MarkupSafe==3.0.3
+SQLAlchemy==2.0.49
+typing_extensions==4.15.0
 Werkzeug==3.1.8
-Flask-SQLAlchemy==3.1.1
